+++ rust/operator-binary/src/controller/mod.rs
@@ -1,1 +1,2 @@
 pub mod dereference;
+pub mod validate;
